+++ src/RestApi/Models.py
@@ -23,7 +23,6 @@
 #
 import os
 import re
-import urllib.request
 import uuid
 import json
 import dateutil
@@ -33,33 +32,7 @@ from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 #from flask_sqlalchemy import SQLAlchemy
 from werkzeug.security import generate_password_hash, check_password_hash
-from flask_oidc import OpenIDConnect
-import jwt
-from jwt import PyJWKClient
 from Config.Config import G_CONFIG
-
-token = "eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICIwazQ1Q2VCS3ZUUFFrR1NocllwcEJEczVTZUxTZGtFaW9IZUllSlZBR1JzIn0.eyJleHAiOjE2ODk4Nzg5MzcsImlhdCI6MTY4OTg3ODYzNywianRpIjoiYzZkMWVjNTgtMDNiNS00YjcwLWI1NTQtNGJjMjZiOWY2Y2U4IiwiaXNzIjoiaHR0cHM6Ly9zc28uZGV2LnRvbGtldmFyYXYuZWtpLmVlL3JlYWxtcy90b2xrZXZhcmF2LWRldiIsImF1ZCI6ImFjY291bnQiLCJzdWIiOiJlZDdkOTFkNS1kMzQ5LTQxYjQtODRjMC0xNGMxYWUwYTgzNjkiLCJ0eXAiOiJCZWFyZXIiLCJhenAiOiJ3ZWIiLCJzZXNzaW9uX3N0YXRlIjoiOGJhNzRlMWUtNTQ5MS00Y2FmLThiODItNGFlNzdkNDk4NTk4IiwiYWNyIjoiMSIsImFsbG93ZWQtb3JpZ2lucyI6WyIqIl0sInJlYWxtX2FjY2VzcyI6eyJyb2xlcyI6WyJvZmZsaW5lX2FjY2VzcyIsImRlZmF1bHQtcm9sZXMtdG9sa2V2YXJhdiBkZXYiLCJ1bWFfYXV0aG9yaXphdGlvbiJdfSwicmVzb3VyY2VfYWNjZXNzIjp7ImFjY291bnQiOnsicm9sZXMiOlsibWFuYWdlLWFjY291bnQiLCJtYW5hZ2UtYWNjb3VudC1saW5rcyIsInZpZXctcHJvZmlsZSJdfX0sInNjb3BlIjoicHJvZmlsZSBlbWFpbCIsInNpZCI6IjhiYTc0ZTFlLTU0OTEtNGNhZi04YjgyLTRhZTc3ZDQ5ODU5OCIsInRvbGtldmFyYXYiOnsicGVyc29uYWxJZGVudGlmaWNhdGlvbkNvZGUiOiI2MDAwMTAxNzcxNiIsInVzZXJJZCI6Ijk5M2JlYjExLTM4NTctNDBmYi1hMmZlLWIxMTQyZGZmY2E0ZiIsImluc3RpdHV0aW9uVXNlcklkIjoiOTkzYmViMTEtMzk5MC00MWZhLTgyYjMtNDIwYmExMDU5ZjJkIiwiZm9yZW5hbWUiOiJkcmFnYW4iLCJzdXJuYW1lIjoiUmlzdG92c2tpIiwic2VsZWN0ZWRJbnN0aXR1dGlvbiI6eyJpZCI6Ijk5M2JlYjExLTE2OWYtNDAzMS05ZjJhLTM3N2JlYjVlNTI4NCIsIm5hbWUiOiJEcmFnYW4ncyBMYW5ndWFnZSBidXJlYXUifSwicHJpdmlsZWdlcyI6WyJBRERfUk9MRSIsIlZJRVdfUk9MRSIsIkVESVRfUk9MRSIsIkRFTEVURV9ST0xFIiwiQUREX1VTRVIiLCJFRElUX1VTRVIiLCJWSUVXX1VTRVIiLCJFWFBPUlRfVVNFUiIsIkFDVElWQVRFX1VTRVIiLCJERUFDVElWQVRFX1VTRVIiLCJBUkNISVZFX1VTRVIiLCJFRElUX1VTRVJfV09SS1RJTUUiLCJFRElUX1VTRVJfVkFDQVRJT04iXX0sImVtYWlsX3ZlcmlmaWVkIjpmYWxzZSwibmFtZSI6IkRyYWdhbiBSaXN0b3Zza2kiLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJlZTYwMDAxMDE3NzE2IiwiZ2l2ZW5fbmFtZSI6IkRyYWdhbiIsImZhbWlseV9uYW1lIjoiUmlzdG92c2tpIn0.QigubLhi7Gf_M2cL9KrfXeAarE1sXtiyt4KAnM1wT1OhGB-rpoKYvg2e3vRjrm_jOe7UmQO4e5Cbwd02vkNIZkjxS370IoZL2du5IKVljNcoScV8o3SHhp221k1N5PEFTO1eriBFFZfNnMjeMAANxDv6Uuf9b0PXIrN_E55rZK7liIApwOD0oIZA-O5CUFaucGv0D3yQ_tcGDyWkoPQkwjMuTD4zHVj2frrJ6AIndBMAjVDrgPvXqH5Cz0Kow72uuWBUQSM4Lb7L504HBpaMTQMztNRuvFPWXu-ub4NaLCbZyzZkgkZEyTVG3I3iW0POlrckRcTQbGjDGZUeDW6TGQ"
-
-def get_jwks_url(issuer_url):
-  well_known_url = issuer_url + "/.well-known/openid-configuration"
-  with urllib.request.urlopen(well_known_url) as response:
-    well_known = json.load(response)
-  if not 'jwks_uri' in well_known:
-    raise Exception('jwks_uri not found in OpenID configuration')
-  return well_known['jwks_uri']
-
-
-def decode_and_validate_token(token):
-  unvalidated = jwt.decode(token, options={"verify_signature": False})
-  jwks_url = get_jwks_url(unvalidated['iss'])
-  jwks_client = jwt.PyJWKClient(jwks_url)
-  header = jwt.get_unverified_header(token)
-  key = jwks_client.get_signing_key(header["kid"]).key
-  return jwt.decode(token, key, [header["alg"]])
-
-
-decoded = decode_and_validate_token(token)
-print(decoded)
 
 
 POSTGRESQL_HOST = G_CONFIG.config['postgresql']['host']
@@ -83,38 +56,6 @@ app.config['PROPAGATE_EXCEPTIONS'] = True
 app.config['JWT_AUTH_HEADER_PREFIX'] = "Bearer"
 
 keycloak_config = G_CONFIG.config['keycloak']
-
-if not os.path.exists('client_secrets.json'):
-  with open('client_secrets.json', 'w') as f:
-    content = {
-        keycloak_config['client_id']: {
-          "issuer": ''.join([keycloak_config['url'],'realms/',keycloak_config['realm']]),
-          "auth_uri": ''.join([keycloak_config['url'],'realms/',keycloak_config['realm'],'/protocol/openid-connect/auth']),
-          "client_id": keycloak_config['client_id'],
-          "client_secret": keycloak_config['client_secret'],
-          "redirect_uris": keycloak_config['redirect_uris'],
-          "userinfo_uri": ''.join([keycloak_config['url'],'realms/',keycloak_config['realm'],'/protocol/openid-connect/userinfo']),
-          "token_uri": ''.join([keycloak_config['url'],'realms/',keycloak_config['realm'],'/protocol/openid-connect/token']),
-          "token_introspection_uri": ''.join([keycloak_config['url'],'realms/',keycloak_config['realm'],'/protocol/openid-connect/token/introspect'])
-        }
-      }
-    f.write(json.dumps(content, indent=2))
-
-app.config.update({
-    'TESTING': True,
-    'DEBUG': True,
-    'OIDC_CLIENT_SECRETS': 'client_secrets.json',
-    'OIDC_ID_TOKEN_COOKIE_SECURE': False,
-    'OIDC_REQUIRE_VERIFIED_EMAIL': False,
-    'OIDC_USER_INFO_ENABLED': True,
-    'OIDC_OPENID_REALM': 'master',
-    'OIDC_SCOPES': ['openid', 'email', 'profile'],
-    'OIDC_INTROSPECTION_AUTH_METHOD': 'client_secret_post',
-    'OIDC_CALLBACK_ROUTE': '/oidc/callback'
-})
-
-oidc = OpenIDConnect(app)
-oidc.init_app(app)
 
 
 # Class to add, update and delete data via SQLALchemy sessions
@@ -160,79 +101,8 @@ class CRUD:
     return d
 
 
-class Users(db.Model):
-  uuid = db.Column(db.Text, primary_key=True)
-  username = db.Column(db.Text)
-  password = db.Column(db.Text)
-  role = db.Column(db.Text, default="user")
-  token_expires = db.Column(db.Boolean, default=True)
-  is_active = db.Column(db.Boolean, default=True)
-  scopes = db.relationship('UserScopes', backref='user')
-  settings = db.relationship('UserSettings', backref='user')
-
-  created = db.Column(db.DateTime, default=datetime.datetime.utcnow)
-
-  # Needed for JWT authentication
-  @property
-  def id(self):
-    return self.uuid
-
-  ADMIN = 'admin'
-
-  def __init__(self, **kwargs):
-    self.uuid = uuid.uuid4()
-    self.update(**kwargs)
-
-  def update(self, **kwargs):
-    for key,value in kwargs.items():
-      if value == '': continue
-      if hasattr(self, key): setattr(self, key, value)
-      if key == 'password': self.set_password(value)
-
-  # Add/update user scope
-  def update_scope(self, **kwargs):
-    if 'id' in kwargs and kwargs['id']:
-      scope = UserScopes.query.get(kwargs['id'])
-      if not scope or scope.user_uuid != self.uuid: return None
-    else:
-      scope = UserScopes(self.uuid)
-
-    for key,value in kwargs.items():
-      if key == "tags":
-        key = "domains" # keep backward compatitbility for now
-      elif not value:
-        continue
-      if re.search('_date$', key):  # convert string to datetime
-        value = dateutil.parser.parse(value)
-      if hasattr(scope, key): setattr(scope, key, value)
-    return scope
-
-  def get_scope(self, id):
-    scope = UserScopes.query.get(id)
-    if not scope or scope.user_uuid != self.uuid:
-      return None
-    return scope
-
-  def delete_scopes(self):
-    UserScopes.query.filter_by(user_uuid = self.uuid).delete()
-
-  def set_password(self, password):
-    self.password = generate_password_hash(password)
-
-  def check_password(self, password):
-    if not self.password: return True
-    if not password: return False
-    return check_password_hash(self.password, password)
-
-  def to_dict(self):
-    d = CRUD.to_dict(self)
-    del d["password"]
-    d["scopes"] = [ s.to_dict() for s in self.scopes]
-    return d
-
 class UserScopes(db.Model):
   id = db.Column(db.Integer, primary_key=True)
-
   user_uuid = db.Column(db.Text, db.ForeignKey('users.uuid'))
 
   # Permission patterns (list of wildcards)
@@ -268,6 +138,34 @@ class UserScopes(db.Model):
     except Exception:
       db.session.rollback()
 
+  # Add/update user scope
+  def update_scope(self, **kwargs):
+    if 'id' in kwargs and kwargs['id']:
+      scope = UserScopes.query.get(kwargs['id'])
+      if not scope or scope.user_uuid != self.uuid: return None
+    else:
+      scope = UserScopes(self.uuid)
+
+    for key,value in kwargs.items():
+      if key == "tags":
+        key = "domains" # keep backward compatitbility for now
+      elif not value:
+        continue
+      if re.search('_date$', key):  # convert string to datetime
+        value = dateutil.parser.parse(value)
+      if hasattr(scope, key): setattr(scope, key, value)
+    return scope
+
+  def get_scope(self, id):
+    scope = UserScopes.query.get(id)
+    if not scope or scope.user_uuid != self.uuid:
+      return None
+    return scope
+
+  def delete_scopes(self):
+    UserScopes.query.filter_by(user_uuid = self.uuid).delete()
+
+
 class UserSettings(db.Model):
   id = db.Column(db.Integer, primary_key=True)
 
@@ -280,6 +178,7 @@ class UserSettings(db.Model):
 
   def to_dict(self):
     return CRUD.to_dict(self)
+
 
 class Tags(db.Model):
    id = db.Column(db.Text, primary_key=True)
@@ -325,7 +224,6 @@ class Tags(db.Model):
        tag = Tags.query.get(tag_id)
        if tag and tag.type != "unspecified": return True
      return False
-
 
    @staticmethod
    def has_public_name(tag_names):
